% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_dt.R
\name{create_dt}
\alias{create_dt}
\title{Create an interactive data table using the package DT}
\usage{
create_dt(data)
}
\arguments{
\item{data}{a tibble or dataframe}
}
\value{
An interactive datatable using the DT package
}
\description{
Create an interactive data table using the package DT
}
\examples{
create_dt(data)
}
