% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mann_kendall_heatmap.R,
%   R/mann_kendall_heatmap_bw.R, R/mann_kendall_reduced_test.R,
%   R/mann_kendall_test.R, R/mk_analysis.R, R/timeseries_plot.R
\docType{data}
\name{date}
\alias{date}
\title{Function to plot mann-kendall trends on a heatmap}
\format{
An object of class \code{NULL} of length 0.

An object of class \code{NULL} of length 0.

An object of class \code{NULL} of length 0.

An object of class \code{NULL} of length 0.

An object of class \code{NULL} of length 0.

An object of class \code{NULL} of length 0.
}
\usage{
date

date

date

date

date

date
}
\arguments{
\item{heatmap_colours}{vector of colours to use for each category/tile}

\item{label_text_size}{size of trend labels}

\item{plot_title}{Text for plot title}

\item{traditional}{perform the standard analysis and not analyse for "Probably" or "Stable" trends}

\item{data}{tibble}

\item{date_size}{size of x-axis date labels}

\item{date_break}{date breaks to be used ("2 weeks")}

\item{date_label}{format of date label (\%b-\%y)}

\item{dates_range}{Range of dates to appear on x-axis}

\item{x_angle}{angle of x axis text}

\item{legend_text_size}{size of the text in the legend}

\item{y_title_size}{size of y_axis label}

\item{y_unit}{unit to display on y_axis heading}

\item{ymin}{Minimum value on y-axis}

\item{ymax}{Maximum value on y-axis}
}
\value{
ggplot heatmap

ggplot heatmap

tibble with trend data

A nested tibble of trends as well as the original nested data

tibble with result and stats

a timeseries plot based on the scale and date range of the dataset
}
\description{
Function to plot mann-kendall trends on a heatmap

Black and White Heatmap for Mann-Kendall Test

Mann_Kendall Function to loop through data set but only lists trends as: Not Significant, Increasing or Decreasing

Mann Kendall Function to loop through entire dataset

Mann_Kendall Test returning test result and stats

Timeseries Plot with standard formatting
}
\examples{
mann_kendall_heatmap(mk_export)
mann_kendall_heatmap_bw(mk_export_reduced)
mann_kendall_reduced_test(mk_export)
mann_kendall_test(data)
mk_analysis(df)
timeseries_plot(data)
}
\keyword{datasets}
