% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mann_kendall_reduced_test.R
\name{mann_kendall_reduced_test}
\alias{mann_kendall_reduced_test}
\title{Mann_Kendall Function to loop through data set but only lists trends as: Not Significant, Increasing or Decreasing}
\usage{
mann_kendall_reduced_test(data)
}
\arguments{
\item{data}{output tibble from data_processor or similar}
}
\value{
tibble with trend data
}
\description{
Mann_Kendall Function to loop through data set but only lists trends as: Not Significant, Increasing or Decreasing
}
\examples{
mann_kendall_reduced_test(mk_export)
}
