% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scale_y_limitval.R
\name{scale_limitval}
\alias{scale_y_limitval}
\alias{scale_x_limitval}
\title{Position scales for continuous data (x & y) with labelled vertical/horizontal
markers}
\usage{
scale_y_limitval(
  marker_values,
  marker_colours = "black",
  marker_labels = marker_values,
  marker_linetypes = 5,
  trans = "identity",
  ...
)

scale_x_limitval(
  marker_values,
  marker_colours = "black",
  marker_labels = marker_values,
  marker_linetypes = 5,
  trans = "identity",
  ...
)
}
\arguments{
\item{marker_values}{Numeric vector of values at which to draw marker lines.}

\item{marker_colours}{Character vector of colours for marker lines. Should be
the same length as \code{marker_values}.}

\item{marker_labels}{Character vector of labels for marker lines. Should be
the same length as \code{marker_values}. Defaults to using the numeric
values given in \code{marker_values}.}

\item{marker_linetypes}{Vector of values for marker linetypes. Should be the
same length as \code{marker_values}. Defaults to dashed lines (2).}

\item{trans}{The name of a transformation object. See
\code{\link[ggplot2:scale_continuous]{ggplot2::scale_x_continuous()}} for more information.}

\item{...}{Other arguments to pass to
\code{\link[ggplot2:scale_continuous]{ggplot2::scale_x_continuous()}}/\code{\link[ggplot2:scale_continuous]{ggplot2::scale_y_continuous()}}.}
}
\value{
A list of ggplot2 components: one marker line per value, plus the
position scale carrying their labels on a secondary axis.
}
\description{
This function is a wrapper around \code{\link[ggplot2:scale_continuous]{ggplot2::scale_x_continuous()}} or
\code{\link[ggplot2:scale_continuous]{ggplot2::scale_y_continuous()}} which allows users to automatically draw
labelled horizontal or vertical markers at specified points. In the air
quality world, this may be particularly useful to display limit values, such
as those defined in legislation or recommended by health authorities.
}
\details{
This function uses \code{\link[ggplot2:sec_axis]{ggplot2::sec_axis()}} to display the line labels. One can
therefore not pass \code{\link[ggplot2:sec_axis]{ggplot2::sec_axis()}} to \code{...}. If users wish to use
\code{\link[ggplot2:sec_axis]{ggplot2::sec_axis()}} it is recommended to use \code{\link[ggplot2:geom_abline]{ggplot2::geom_abline()}} to
draw lines manually.

Unlike other "scale" functions the order in which this function is added to
the \code{\link[ggplot2:ggplot]{ggplot2::ggplot()}} object matters. Adding \code{scale_*_limitval()}
\emph{after} \code{\link[ggplot2:geom_path]{ggplot2::geom_line()}} will draw the line markers on top of the trend
line, whereas adding it \emph{before} will draw the markers below.
}
\examples{
library(ggplot2)

ggplot(gRs_data, aes(x = date, y = concentration)) +
  geom_point() +
  scale_y_limitval(c(10, 50), marker_labels = c("Guideline", "Trigger"))

ggplot(gRs_data, aes(x = concentration, y = chem_name)) +
  geom_point() +
  scale_x_limitval(10, marker_colours = "red")

}
