% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zero_row_tibble.R
\name{zero_row_tibble}
\alias{zero_row_tibble}
\title{Build a zero row tibble with names}
\usage{
zero_row_tibble(names, base = FALSE)
}
\arguments{
\item{names}{vector of names for the tibble}

\item{base}{should a standard data.frame be produced rather than a tibble?}
}
\value{
Tibble with column names and zero rows
}
\description{
Build a zero row tibble with names
}
\examples{
zero_row_tibble(c("A", "B", "C"))
}
