% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_processor.R
\name{data_processor}
\alias{data_processor}
\title{data_processor}
\usage{
data_processor(myfile_path, sheet_pattern = "Chem")
}
\arguments{
\item{myfile_path}{file path to data}

\item{sheet_pattern}{matching the excel sheet name for new or old esdat formats}
}
\value{
uploaded file from esdat
}
\description{
data_processor
}
\examples{
\dontrun{
data_processor('my_file_path')
}
}
