% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_processor.R
\name{data_processor}
\alias{data_processor}
\title{Data Processor}
\usage{
data_processor(myfile_path)
}
\arguments{
\item{myfile_path}{path to excel file}
}
\value{
tible of processed data
}
\description{
Data Processor
}
\examples{
data_processor("sw_data.xlsx")
}
