% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/establish_plotting_variables.R
\name{establish_plotting_variables}
\alias{establish_plotting_variables}
\title{Establish parameters for automated plots}
\usage{
establish_plotting_variables(data)
}
\arguments{
\item{=esdat}{data frame imported from the data_processor function}
}
\value{
lists and vectors for plots
}
\description{
Establish parameters for automated plots
}
\examples{
anlaytes, date_range, zones, locations and colours. All lists and vectors used for plotting in plot functions
}
