% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/half_lor.R
\name{half_lor}
\alias{half_lor}
\title{Half concentrations that are < LOR}
\usage{
half_lor(data)
}
\arguments{
\item{tibble}{with columns "prefix" and "concentration". Prefix must be either "<" or "="}
}
\value{
tibble
}
\description{
Half concentrations that are < LOR
}
\examples{
<4 becomes =2
}
