% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mann_kendall_heatmap.R
\name{mann_kendall_heatmap}
\alias{mann_kendall_heatmap}
\title{Function to plot mann-kendall trends on a heatmap}
\usage{
mann_kendall_heatmap(
  data,
  label_text_size = 2.8,
  plot_title = "Mann-Kendall Trend Analysis"
)
}
\arguments{
\item{data}{tibble from the mann_kendall_test function}

\item{plot_title}{Title Text}

\item{text_size}{size of the label text}
}
\value{
ggplot heatmap
}
\description{
Function to plot mann-kendall trends on a heatmap
}
\examples{
mann_kendall_heatmap(mk_export)
}
