% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mann_kendall_heatmap.R
\name{mann_kendall_heatmap}
\alias{mann_kendall_heatmap}
\title{Function to plot mann-kendall trends on a heatmap}
\usage{
mann_kendall_heatmap(
  data,
  label_text_size = 3.2,
  plot_title = "Mann-Kendall Trend Analysis",
  heatmap_colours = NULL,
  width = 20
)
}
\arguments{
\item{data}{tibble from the mann_kendall_test function}

\item{label_text_size}{size of the label text}

\item{plot_title}{Title Text}

\item{heatmap_colours}{vector of colours to use for each category/tile
http://127.0.0.1:12443/graphics/plot_zoom_png?width=1920&height=1017}
}
\value{
ggplot heatmap
}
\description{
Function to plot mann-kendall trends on a heatmap
}
\examples{
mann_kendall_heatmap(mk_export)
}
