% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mann_kendall_heatmap_bw.R
\name{mann_kendall_heatmap_bw}
\alias{mann_kendall_heatmap_bw}
\title{Black and White Heatmap for Mann-Kendall Test}
\usage{
mann_kendall_heatmap_bw(
  data,
  label_text_size = 3.5,
  plot_title = "Mann-Kendall Trend Analysis",
  width = 20
)
}
\arguments{
\item{data}{tibble exported from mann_kendall_reduced_test}

\item{label_text_size}{size of trend labels}

\item{plot_title}{Text for plot title}
}
\value{
ggplot heatmap
}
\description{
Black and White Heatmap for Mann-Kendall Test
}
\examples{
mann_kendall_heatmap_bw(mk_export_reduced)
}
