% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mann_kendall_test.R
\name{mann_kendall_test}
\alias{mann_kendall_test}
\title{Mann Kendall Function to loop through entire dataset}
\usage{
mann_kendall_test(data)
}
\arguments{
\item{data}{tibble processed with the data_processor function}
}
\value{
tibble of increasing trends called mk.export
}
\description{
Mann Kendall Function to loop through entire dataset
}
\examples{
mann_kendall_test(data)
}
