% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mann_kendall_test.R
\name{mann_kendall_test}
\alias{mann_kendall_test}
\title{Mann Kendall Function to loop through entire dataset}
\usage{
mann_kendall_test(data, traditional = FALSE)
}
\arguments{
\item{data}{tibble processed with the data_processor function}

\item{traditional}{perform the standard analysis and not analyse for "Probably" or "Stable" trends}
}
\value{
A nested tibble of trends as well as the original nested data
}
\description{
Mann Kendall Function to loop through entire dataset
}
\examples{
mann_kendall_test(data)
}
