% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mk_analysis.R
\name{mk_analysis}
\alias{mk_analysis}
\title{Mann_Kendall Test returning test result and stats}
\usage{
mk_analysis(data)
}
\arguments{
\item{data}{filtered tibble with column of "concentration"}
}
\value{
tibble with result and stats
}
\description{
Mann_Kendall Test returning test result and stats
}
\examples{
mk_analysis(df)
}
