% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_by_analyte.R
\name{plot_by_analyte}
\alias{plot_by_analyte}
\title{Plots a timeseries_plot for each analyte and saves to designated location}
\usage{
plot_by_analyte(data, save_path = NULL, smooth = FALSE, linear_smooth = FALSE)
}
\arguments{
\item{data}{tibble of concentration data}

\item{save_path}{path to directory where plots will be saved}

\item{smooth}{Apply a geom_smooth to the plot}

\item{linear_smooth}{Apply a linear trendline to the plot}

\item{...}{
  Arguments passed on to \code{\link[=timeseries_plot]{timeseries_plot}}
  \describe{
    \item{\code{date_size}}{size of x-axis date labels}
    \item{\code{date_break}}{date breaks to be used ("2 weeks")}
    \item{\code{date_label}}{format of date label (\%b-\%y)}
    \item{\code{x_angle}}{angle of x axis text}
    \item{\code{legend_text_size}}{size of the text in the legend}
  }}
}
\value{
jpeg of timeseries plots
}
\description{
Plots a timeseries_plot for each analyte and saves to designated location
}
\examples{
plot_by_analyte("my_project/figures")
}
