% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary_stats.R
\name{summary_stats}
\alias{summary_stats}
\title{Summary statistics table}
\usage{
summary_stats(data, save_path = NULL)
}
\arguments{
\item{data}{tibble from data_processor}

\item{save_path}{path to save summary tables as .xlsx files}
}
\value{
tibbles and csv files
}
\description{
Summary statistics table
}
\examples{
summary_stats(df, save_path = "users/project/stats")
}
